#549 add mixin utility to C++, implement system as mixin
