+++ include/flecs/addons/cpp/addons/system/builder_i.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "query_i.hpp"
+#include "../../builders/query_i.hpp"
 
 namespace flecs 
 {
